fix:留言修复，数据库修复中，外部依赖组件去除

+++ myblog-server/src/main/java/com/yikolemon/pojo/Like.java
@@ -5,6 +5,7 @@ import lombok.Data;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+import java.util.concurrent.atomic.AtomicInteger;
 
 @Data                              //get，set
 @NoArgsConstructor                 //无参构造
@@ -13,6 +14,6 @@ public class Like implements Serializable {
 
     private int like;
 
-    private int blogId;
+    private long blogId;
 
 }
